+++ packages/emd-basic-money/src/core.js
@@ -1,5 +1,6 @@
 import { compose } from '@stone-payments/emd-helpers';
 import { withComponent } from '@stone-payments/emd-hocs';
+import { LitElement } from '@stone-payments/lit-element';
 
 import { MoneyController } from './component/MoneyController.js';
 import { MoneyView } from './component/MoneyView.js';
@@ -7,7 +8,7 @@ import { MoneyView } from './component/MoneyView.js';
 const Money = compose(
   MoneyController,
   withComponent
-)(HTMLElement);
+)(LitElement);
 
 Money.views = MoneyView;