+++ packages/emd-basic-table/src/component/TableView.js
@@ -1,58 +1,69 @@
 import { html } from '@stone-payments/lit-element';
-import { APPEARANCE_MAP } from '../constants/appearance.js';
 
-const getStypeProperty = (styles, index) => propName =>
-  `${APPEARANCE_MAP[propName]}: ${styles[index][propName]}; `;
-
-const getCellStyle = (styles, index) => Object.keys(APPEARANCE_MAP)
-  .map(getStypeProperty(styles, index))
-  .join('')
-  .trim();
-
-const getTableClass = (clickablerows, expandedbody) => {
+const getTableClass = (clickablerows, expandedbody, view) => {
   let result = 'emd-table__wrapper ';
   result += clickablerows ? 'emd-table__wrapper_clickable ' : '';
   result += expandedbody ? 'emd-table__wrapper_expanded ' : '';
-
+  result += view ? `emd-table__wrapper_view_${view} ` : '';
   return result.trim();
 };
 
 export const TableView = ({
-  headerRow,
-  headerStyles,
-  matrix,
-  matrixStyles,
+  rows = [],
+  wrappedTitles = [],
+  titles = [],
+  handleRowClick,
+  dispatchCustomEvent,
+  getHeaderAdapter,
+  stringifyHeaderAppearance,
+  stringifyHeaderCellAppearance,
+  getRowAdapter,
+  stringifyRowAppearance,
+  stringifyRowCellAppearance,
   clickablerows,
   expandedbody,
-  handleRowClick
+  view
 }) => html`
   <style>
     @import url("emd-basic-table/src/component/Table.css")
   </style>
-  <table class="${getTableClass(clickablerows, expandedbody)}">
-    ${headerRow.length > 0 ? html`
+  <table class="${getTableClass(clickablerows, expandedbody, view)}">
+    ${wrappedTitles.length > 0 ? html`
       <thead class="emd-table__header">
-        <tr class="emd-table__row">
-          ${headerRow.map((cell, cellIndex) => html`
-            <th
-              class="emd-table__cell"
-              style="${getCellStyle(headerStyles, cellIndex)}">
-              ${cell}
-            </th>
-          `)}
-        </tr>
+        ${wrappedTitles.map((row, rowIndex) => html`
+          <tr class="emd-table__row">
+            ${getHeaderAdapter(row, rowIndex)(row, rowIndex, dispatchCustomEvent(rowIndex)).map((cell, cellIndex, cellArray) => html`
+              <th
+                data-label="${row[cellIndex] || ''}"
+                style="${stringifyHeaderAppearance(cellArray.length)[cellIndex]}"
+                class="emd-table__cell">
+                <span
+                  class="emd-table__cell-text"
+                  style="${stringifyHeaderCellAppearance(cellArray.length)[cellIndex]}">
+                  ${cell}
+                </span>
+              </th>
+            `)}
+          </tr>
+        `)}
       </thead>
     ` : ''}
-    ${matrix.length > 0 ? html`
+    ${rows.length > 0 ? html`
       <tbody class="emd-table__body">
-        ${matrix.map((row, rowIndex) => html`
+        ${rows.map((row, rowIndex) => html`
           <tr class="emd-table__row">
-            ${row.map((cell, cellIndex) => html`
+            ${getRowAdapter(row, rowIndex)(row, rowIndex, dispatchCustomEvent(rowIndex)).map((cell, cellIndex, cellArray) => html`
               <td
+                data-label="${titles[cellIndex] || ''}"
                 @click="${handleRowClick(rowIndex)}"
-                class="emd-table__cell"
-                style="${getCellStyle(matrixStyles[rowIndex], cellIndex)}">
-                ${cell}
+                style="${stringifyRowAppearance(row, rowIndex, cellArray.length)[cellIndex]}"
+                class="emd-table__cell">
+                <span class="emd-table__cell-overlay"></span>
+                <span
+                  class="emd-table__cell-text"
+                  style="${stringifyRowCellAppearance(row, rowIndex, cellArray.length)[cellIndex]}">
+                  ${cell}
+                </span>
               </td>
             `)}
           </tr>