+++ packages/emd-basic-brand-icon/CHANGELOG.md
@@ -1,3 +1,23 @@
+# [@stone-payments/emd-basic-brand-icon-v1.1.0](https://github.com/stone-payments/emerald-web-framework/compare/@stone-payments/emd-basic-brand-icon-v1.0.3...@stone-payments/emd-basic-brand-icon-v1.1.0) (2019-09-09)
+
+
+### Bug Fixes
+
+* Correct example page ([9316afe](https://github.com/stone-payments/emerald-web-framework/commit/9316afe))
+* Fix bug on old Edge browsers ([b6dba60](https://github.com/stone-payments/emerald-web-framework/commit/b6dba60))
+
+
+### Features
+
+* Make brand icons accept accents on params ([b5d9770](https://github.com/stone-payments/emerald-web-framework/commit/b5d9770))
+
+# [@stone-payments/emd-basic-brand-icon-v1.0.3](https://github.com/stone-payments/emerald-web-framework/compare/@stone-payments/emd-basic-brand-icon-v1.0.2...@stone-payments/emd-basic-brand-icon-v1.0.3) (2019-09-05)
+
+
+### Bug Fixes
+
+* Fix component on IE Edge ([8ae0227](https://github.com/stone-payments/emerald-web-framework/commit/8ae0227))
+
 # [@stone-payments/emd-basic-brand-icon-v1.0.2](https://github.com/stone-payments/emerald-web-framework/compare/@stone-payments/emd-basic-brand-icon-v1.0.1...@stone-payments/emd-basic-brand-icon-v1.0.2) (2019-09-05)